requirements.txt
passlib[bcrypt]>=1.7.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
fastapi==0.115.0
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==1.51.0
packaging==25.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.38.6
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
